Added api layer #3

+++ internal/repository/user/model/user.go
@@ -7,7 +7,7 @@ type User struct {
 	Name      string
 	Email     string
 	Password  string
-	Role      int64
+	Role      int32
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
